vandelay_search/pyproject.toml
[project]
name = "vandelay-search"
version = "1.0.0"
description = "Vandelay Search Agent - FSI GraphRAG with Google ADK"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-adk>=0.1.0",
    "litellm>=1.0.0",
    "neo4j>=5.0.0",
    "httpx>=0.24.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.20.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["vandelay_search"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]
